packages/dynamicspawner/requirements.txt
kubernetes_asyncio
jupyterhub
jupyterhub-kubespawner
